chore: Better directory and class structure (#609)
