+++ src/pytest_html/selfcontained_report.py
@@ -1,0 +1,36 @@
+import base64
+import binascii
+import warnings
+
+from pytest_html.basereport import BaseReport
+
+
+class SelfContainedReport(BaseReport):
+    def __init__(self, report_path, config):
+        super().__init__(report_path, config)
+
+    @property
+    def css(self):
+        return self._css
+
+    def _data_content(self, content, mime_type, *args, **kwargs):
+        charset = "utf-8"
+        data = base64.b64encode(content.encode(charset)).decode(charset)
+        return f"data:{mime_type};charset={charset};base64,{data}"
+
+    def _media_content(self, content, mime_type, *args, **kwargs):
+        try:
+            # test if content is base64
+            base64.b64decode(content.encode("utf-8"), validate=True)
+            return f"data:{mime_type};base64,{content}"
+        except binascii.Error:
+            # if not base64, issue warning and just return as it's a file or link
+            warnings.warn(
+                "Self-contained HTML report "
+                "includes link to external "
+                f"resource: {content}"
+            )
+            return content
+
+    def _generate_report(self, *args, **kwargs):
+        super()._generate_report(self_contained=True)
